Table Products is added
