+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "build": "tsc",
-    "start": "node dist/index.js",
+    "start": " npm run build && node dist/index.js",
     "clean": "rm -rf build",
     "dev": "nodemon"
   },
