+++ src/Models/index.ts
@@ -1,0 +1,4 @@
+import { User } from "./user";
+import { Product } from "./product";
+
+export { User, Product };
